ZSET_SCORE_MIN/MAX are really should be double (#4)

+++ include/blackwidow/blackwidow.h
@@ -23,8 +23,8 @@
 
 namespace blackwidow {
 
-const int64_t ZSET_SCORE_MAX = std::numeric_limits<double>::max();
-const int64_t ZSET_SCORE_MIN = std::numeric_limits<double>::lowest();
+const double ZSET_SCORE_MAX = std::numeric_limits<double>::max();
+const double ZSET_SCORE_MIN = std::numeric_limits<double>::lowest();
 
 const std::string USAGE_TYPE_ALL = "all";
 const std::string USAGE_TYPE_NEMO = "nemo";
